backend/pyproject.toml
[tool.poetry]
name = "rag-pipeline"
version = "0.1.0"
description = ""
authors = ["imnotdev25 <85677268+imnotdev25@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
pypdf2 = "^3.0.1"
sqlmodel = "^0.0.22"
psycopg2-binary = "^2.9.9"
pgvector = "^0.3.4"
pydantic-settings = "^2.5.2"
uvicorn = "^0.31.0"
python-multipart = "^0.0.12"
python-docx = "^1.1.2"
python-pptx = "^1.0.2"
bs4 = "^0.0.2"
pandas = "^2.2.3"
tiktoken = "^0.8.0"
langchain-experimental = "^0.3.2"
logfire = {extras = ["fastapi", "httpx", "openai"], version = "^2.6.2"}
aiofiles = "^24.1.0"
openai = "^1.57.2"
asyncpg = "^0.30.0"

[tool.poetry.dev-dependencies]
ruff = "^0.8.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

